instance_manager: Move out Metrics to its own module
